@@ -1,13 +1,7 @@
-﻿using Newtonsoft.Json;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Text.Json;
-using System.Threading;
-using System.Threading.Tasks;
+using System.Net;
 using TlsClient.Core.Helpers;
-using TlsClient.Core.Helpers.Natives;
 using TlsClient.Core.Helpers.Wrappers;
 using TlsClient.Core.Models.Entities;
 using TlsClient.Core.Models.Requests;
@@ -18,24 +12,32 @@ namespace TlsClient.Core
 {
     public class TlsClient : IDisposable
     {
-        private IntPtr LoadedLibrary { get; set; } = IntPtr.Zero;
+        private bool _disposed;
+
         public TlsClientOptions Options { get; set; }
         public Dictionary<string, List<string>> DefaultHeaders => Options.DefaultHeaders;
-        private TlsClientWrapper _wrapper { get; set; }
+        public bool IsDisposed => _disposed;
+
         public TlsClient(TlsClientOptions options)
         {
             Options = options ?? throw new ArgumentNullException(nameof(options));
-            LoadedLibrary = NativeLoader.LoadNativeAssembly(options.LibraryPath);
-            _wrapper = new TlsClientWrapper(LoadedLibrary);
         }
 
         public TlsClient() : this(new TlsClientOptions(TlsClientIdentifier.Chrome132, "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/132.0.0.0 Safari/537.36 OPR/117.0.0.0")) { }
-        public async Task<Response> RequestAsync(Request request, CancellationToken cancellationToken = default)
+
+        public static void Initialize(string? libraryPath) => TlsClientWrapper.Initialize(libraryPath);
+
+        public Response Request(Request request)
         {
-            if (request.RequestUrl == null)
+            if (request is null) throw new ArgumentNullException(nameof(request));
+            if (string.IsNullOrWhiteSpace(request.RequestUrl))
+                throw new ArgumentException("RequestUrl cannot be null or empty.", nameof(request));
+
+            if ((request.TlsClientIdentifier == null && Options.TlsClientIdentifier == null) && (request.CustomTlsClient == null && Options.CustomTlsClient == null))
             {
-                throw new ArgumentNullException(nameof(request.RequestUrl));
+                throw new ArgumentException("Either TlsClientIdentifier or CustomTlsClient must be set in Options or in Request.");
             }
+
             request.TlsClientIdentifier ??= Options.TlsClientIdentifier;
             request.SessionId ??= Options.SessionID;
             request.TimeoutMilliseconds ??= (int)Options.Timeout.TotalMilliseconds;
@@ -49,8 +51,14 @@ namespace TlsClient.Core
             request.WithDefaultCookieJar ??= Options.WithDefaultCookieJar;
             request.WithoutCookieJar ??= Options.WithoutCookieJar;
             request.CustomTlsClient ??= Options.CustomTlsClient;
-           
-            if(request.CustomTlsClient != null) // If CustomTlsClient is not null, TlsClientIdentifier should be null
+            request.CatchPanics ??= Options.CatchPanics;
+            request.CertificatePinningHosts ??= Options.CertificatePinningHosts;
+            request.ForceHttp1 ??= Options.ForceHttp1;
+            request.WithRandomTLSExtensionOrder ??= Options.WithRandomTLSExtensionOrder;
+            request.HeaderOrder ??= Options.HeaderOrder;
+            request.ConnectHeaders ??= Options.ConnectHeaders;
+
+            if (request.CustomTlsClient != null) // If CustomTlsClient is not null, TlsClientIdentifier should be null
             {
                 request.TlsClientIdentifier = default!;
             }
@@ -69,14 +77,16 @@ namespace TlsClient.Core
             }
 
             // Override if has Host header
-            request.RequestHostOverride ??= request.Headers.ContainsKey("Host") ? request.Headers["Host"] : null;
+            if (request.Headers != null && request.Headers.TryGetValue("Host", out var host) && !string.IsNullOrWhiteSpace(host))
+                request.RequestHostOverride ??= host;
 
             Response response;
 
             try
             {
-                var rawResponse = await _wrapper.RequestAsync(RequestHelpers.Prepare(request), cancellationToken);
-                response = JsonConvert.DeserializeObject<Response>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
+                var payload = RequestHelpers.Prepare(request);
+                var rawResponse = TlsClientWrapper.Request(payload);
+                response = RequestHelpers.ConvertObject<Response>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
             }
             catch(Exception err)
             {
@@ -88,72 +98,84 @@ namespace TlsClient.Core
             }
 
             if(!string.IsNullOrEmpty(response.Id))
-                await _wrapper.FreeMemoryAsync(response.Id, cancellationToken);
+                TlsClientWrapper.FreeMemory(response.Id);
+
+            if (response.Status == 0 && response.Body.Contains("Client.Timeout exceeded"))
+            {
+                response = new Response()
+                {
+                    Body = "Timeout",
+                    Status = HttpStatusCode.RequestTimeout,
+                };
+            }
 
             return response;
         }
 
-        public async Task<GetCookiesFromSessionResponse> GetCookiesAsync(string url, CancellationToken cancellationToken = default)
+        public GetCookiesFromSessionResponse GetCookies(string url)
         {
-            if (string.IsNullOrEmpty(url))
-            {
-                throw new ArgumentNullException(nameof(url));
-            }
+            if (string.IsNullOrWhiteSpace(url))
+                throw new ArgumentException("url cannot be null or empty.", nameof(url));
 
             var payload = new GetCookiesFromSessionRequest()
             {
                 SessionID = Options.SessionID,
                 Url = url,
             };
-
-            var rawResponse = await _wrapper.GetCookiesFromSessionAsync(RequestHelpers.Prepare(payload), cancellationToken);
-            return JsonConvert.DeserializeObject<GetCookiesFromSessionResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
+            var rawResponse = TlsClientWrapper.GetCookiesFromSession(RequestHelpers.Prepare(payload));
+            return RequestHelpers.ConvertObject<GetCookiesFromSessionResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
         }
 
-        public async Task<DestroyResponse> DestroyAsync(CancellationToken cancellationToken = default)
+        public DestroyResponse Destroy()
         {
             var payload = new DestroyRequest()
             {
                 SessionID = Options.SessionID,
             };
-
-            var rawResponse = await _wrapper.DestroySessionAsync(RequestHelpers.Prepare(payload), cancellationToken);
-            return JsonConvert.DeserializeObject<DestroyResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
+            
+            var rawResponse = TlsClientWrapper.DestroySession(RequestHelpers.Prepare(payload));
+            return RequestHelpers.ConvertObject<DestroyResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
         }
 
-        public async Task<DestroyResponse> DestroyAllAsync(CancellationToken cancellationToken = default)
+        public static DestroyResponse DestroyAll()
         {
-            var rawResponse = await _wrapper.DestroyAllAsync(cancellationToken);
-            return JsonConvert.DeserializeObject<DestroyResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
+            var rawResponse = TlsClientWrapper.DestroyAll();
+            return RequestHelpers.ConvertObject<DestroyResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
         }
 
-        public async Task<GetCookiesFromSessionResponse> AddCookiesAsync(string url, List<TlsClientCookie> cookies, CancellationToken cancellationToken = default)
+        public GetCookiesFromSessionResponse AddCookies(string url, List<TlsClientCookie> cookies)
         {
             if (string.IsNullOrEmpty(url))
             {
                 throw new ArgumentNullException(nameof(url));
             }
 
-            var payload = new AddCookiesToSessionRequest()
+            var payload = new AddCookiesToSessionRequest
             {
                 SessionID = Options.SessionID,
                 Url = url,
                 Cookies = cookies,
             };
-            var rawResponse = await _wrapper.AddCookiesToSessionAsync(RequestHelpers.Prepare(payload), cancellationToken);
-            return JsonConvert.DeserializeObject<GetCookiesFromSessionResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
+            var rawResponse = TlsClientWrapper.AddCookiesToSession(RequestHelpers.Prepare(payload));
+            return RequestHelpers.ConvertObject<GetCookiesFromSessionResponse>(rawResponse) ?? throw new Exception("Response is null, can't convert object from json.");
         }
-  
+
 
         public void Dispose()
         {
-            if(LoadedLibrary == IntPtr.Zero) return;
-
-            DestroyAsync().GetAwaiter().GetResult();
-            _wrapper.Dispose();
-
-            LoadedLibrary = IntPtr.Zero;
+            Dispose(true);
             GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (_disposed) return;
+            if (disposing)
+            {
+                try { Destroy(); }
+                catch { }
+            }
+            _disposed = true;
         }
     }
 }
